[fluentd-elasticsearch] Fix jobLabel in ServiceMonitor (#312)
The value of `jobLabel` in a `ServiceMonitor` defines the label of the
scraped service whose value should be used as `job_name` by Prometheus.
However, it looks like in the chart it was being used by mistake as if it
was meant to be the actual value one wanted to use for the job. Because of
this, scraped metrics would end up with, say,
`fluentd-elasticsearch-metrics` (the name of the scraped `Service`) instead
of `fluentd-elasticsearch`.

This was particularly tricky since it would interact badly with the
provided `PrometheusRules`. For example, `FluentdNodeDown`, that would
normally be triggered when there are no fluentd pods running, gets defined as
`up{job="fluentd-elasticsearch"} == 0`, but because there were no
samples generated for that job, it would remain forever green!

Signed-off-by: Daniel Gorin <jcpetruzza@gmail.com>

Co-authored-by: André Bauer <monotek@users.noreply.github.com>